--- launch/perception.yaml
@@ -1,101 +1,0 @@
-description: "Run the yielding trajectory planner on the real vehicle with real perception"
-mode: hardware
-vehicle_interface: gem_hardware.GEMHardwareInterface
-mission_execution: StandardExecutor
-
-# Recovery behavior after a component failure
-recovery: 
-    planning: 
-        trajectory_tracking : recovery.StopTrajectoryTracker
-
-# Driving behavior for the GEM vehicle.  Runs real pedestrian perception, yield planner, but does not send commands to real vehicle.
-drive: 
-    perception:
-        state_estimation : GNSSStateEstimator
-        agent_detection : pedestrian_detection.PedestrianDetector2D
-        perception_normalization : StandardPerceptionNormalizer
-    planning:
-        relations_estimation: pedestrian_yield_logic.PedestrianYielder
-        route_planning:
-            type: StaticRoutePlanner
-            args: [!relative_path '../GEMstack/knowledge/routes/forward_15m.csv','start']
-        motion_planning: longitudinal_planning.YieldTrajectoryPlanner
-        trajectory_tracking:
-            type: pure_pursuit.PurePursuitTrajectoryTracker
-            print: False
-
-
-log:
-    # Specify the top-level folder to save the log files.  Default is 'logs'
-    #folder : 'logs' 
-    # If prefix is specified, then the log folder will be named with the prefix followed by the date and time. Default no prefix
-    #prefix : 'fixed_route_'  
-    # If suffix is specified, then logs will output to folder/prefix+suffix. Default uses date and time as the suffix
-    #suffix : 'test3'  
-    # Specify which ros topics to record to vehicle.bag. Default records nothing. This records the "standard" ROS topics.
-    ros_topics : 
-    # Specify options to pass to rosbag record. Default is no options.
-    #rosbag_options : '--split --size=1024' 
-    # If True, then record all readings / commands of the vehicle interface. Default False
-    vehicle_interface : True
-    # Specify which components to record to behavior.json. Default records nothing
-    components : ['state_estimation','agent_detection','motion_planning']
-    # Specify which components of state to record to state.json. Default records nothing 
-    #state: ['all']
-    # Specify the rate in Hz at which to record state to state.json. Default records at the pipeline's rate
-    #state_rate: 10
-replay:  # Add items here to set certain topics / inputs to be replayed from logs
-    # Specify which log folder to replay from
-    log: 
-    ros_topics : []
-    components : []
-
-#usually can keep this constant
-computation_graph: !include "../GEMstack/knowledge/defaults/computation_graph.yaml"
-
-variants:
-    detector_only:
-        run:
-            description: "Run the pedestrian detection code"
-            drive:
-                planning:
-                    trajectory_tracking:
-            visualization: !include "mpl_visualization.yaml"
-
-    real_sim:
-        run:
-            description: "Run the pedestrian detection code with real detection and fake simulation"
-            mode: hardware
-            vehicle_interface:
-                type: gem_mixed.GEMRealSensorsWithSimMotionInterface
-                args:
-                    scene: !relative_path '../scenes/xyhead_demo.yaml'
-            mission_execution: StandardExecutor
-            require_engaged: False
-            visualization: !include "klampt_visualization.yaml"
-            drive: 
-                perception:
-                    state_estimation : OmniscientStateEstimator
-
-
-    fake_real:
-        run:
-            description: "Run the yielding trajectory planner on the real vehicle with faked perception"
-            drive:
-                perception:
-                    agent_detection : pedestrian_detection.FakePedestrianDetector2D
-
-    fake_sim:
-        run:
-            description: "Run the yielding trajectory planner in simulation with faked perception"
-            mode: simulation
-            vehicle_interface:
-                type: gem_simulator.GEMDoubleIntegratorSimulationInterface
-                args:
-                    scene: !relative_path '../scenes/xyhead_demo.yaml'
-            visualization: !include "klampt_visualization.yaml"
-            drive: 
-                perception:
-                    agent_detection : pedestrian_detection.FakePedestrianDetector2D
-                    #agent_detection : OmniscientAgentDetector  #this option reads agents from the simulator
-                    state_estimation : OmniscientStateEstimator
